--- DTO/src/dto/api/DTOMenu3ForUi.java
@@ -1,8 +1,0 @@
-package dto.api;
-
-import dto.definition.property.api.PropertyDefinitionDTO;
-import java.util.List;
-
-public interface DTOMenu3ForUi {
-    List<PropertyDefinitionDTO> getEnvironmentVars();
-}
