+++ DTO/src/dto/api/DTOMenu3ForUiContext.java
@@ -1,0 +1,5 @@
+package dto.api;
+
+public interface DTOMenu3ForUiContext {
+    Context getContext();
+}
