+++ DTO/src/dto/api/DTOMenu3ForUiEVD.java
@@ -1,0 +1,8 @@
+package dto.api;
+
+import dto.definition.property.definition.api.PropertyDefinitionDTO;
+import java.util.List;
+
+public interface DTOMenu3ForUiEVD {
+    List<PropertyDefinitionDTO> getEnvironmentVars();
+}
