+++ DTO/src/dto/api/DTOMenu3ForUiEVI.java
@@ -1,0 +1,10 @@
+package dto.api;
+
+import dto.definition.property.definition.api.PropertyDefinitionDTO;
+import dto.definition.property.instance.api.PropertyInstanceDTO;
+
+import java.util.List;
+
+public interface DTOMenu3ForUiEVI {
+    List<PropertyInstanceDTO> getEnvironmentVars();
+}
