+++ DTO/src/dto/definition/entity/api/EntityDefinitionDTO.java
@@ -1,7 +1,7 @@
 package dto.definition.entity.api;
 
 
-import dto.definition.property.api.PropertyDefinitionDTO;
+import dto.definition.property.definition.api.PropertyDefinitionDTO;
 
 import java.util.List;
