+++ DTO/src/dto/definition/environment/variable/api/EnvVariablesDefinitionDTOManager.java
@@ -1,6 +1,6 @@
 package dto.definition.environment.variable.api;
 
-import dto.definition.property.api.PropertyDefinitionDTO;
+import dto.definition.property.definition.api.PropertyDefinitionDTO;
 import java.util.Collection;
 
 public interface EnvVariablesDefinitionDTOManager {
