+++ DTO/src/dto/definition/property/definition/value/generator/api/ValueGenerator.java
@@ -1,0 +1,5 @@
+package dto.definition.property.definition.value.generator.api;
+
+public interface ValueGenerator<T> {
+    T generateValue();
+}
