+++ DTO/src/dto/definition/property/definition/value/generator/impl/random/impl/bool/RandomBooleanValueGenerator.java
@@ -1,0 +1,12 @@
+package dto.definition.property.definition.value.generator.impl.random.impl.bool;
+
+
+import dto.definition.property.definition.value.generator.impl.random.api.AbstractRandomValueGenerator;
+
+public class RandomBooleanValueGenerator extends AbstractRandomValueGenerator<Boolean> {
+
+    @Override
+    public Boolean generateValue() {
+        return random.nextDouble() < 0.5;
+    }
+}
