+++ DTO/src/dto/definition/property/instance/api/PropertyInstanceDTO.java
@@ -1,0 +1,10 @@
+package dto.definition.property.instance.api;
+
+
+import dto.definition.property.definition.api.PropertyDefinitionDTO;
+
+public interface PropertyInstanceDTO {
+    PropertyDefinitionDTO getPropertyDefinition();
+    Object getValue();
+
+}
