--- DTO/src/dto/definition/property/value/generator/api/ValueGenerator.java
@@ -1,5 +1,0 @@
-package dto.definition.property.value.generator.api;
-
-public interface ValueGenerator<T> {
-    T generateValue();
-}
