--- DTO/src/dto/definition/property/value/generator/impl/random/impl/bool/RandomBooleanValueGenerator.java
@@ -1,12 +1,0 @@
-package dto.definition.property.value.generator.impl.random.impl.bool;
-
-
-import dto.definition.property.value.generator.impl.random.api.AbstractRandomValueGenerator;
-
-public class RandomBooleanValueGenerator extends AbstractRandomValueGenerator<Boolean> {
-
-    @Override
-    public Boolean generateValue() {
-        return random.nextDouble() < 0.5;
-    }
-}
