--- DTO/src/dto/definition/property/value/generator/impl/random/impl/numeric/RandomFloatGenerator.java
@@ -1,13 +1,0 @@
-package dto.definition.property.value.generator.impl.random.impl.numeric;
-
-public class RandomFloatGenerator extends AbstractNumericRandomGenerator<Float> {
-
-    public RandomFloatGenerator(Float from, Float to) {
-        super(from, to);
-    }
-
-    @Override
-    public Float generateValue() {
-        return (float)(to-from) * random.nextFloat();
-    }
-}
