--- DTO/src/dto/definition/property/value/generator/impl/random/impl/numeric/RandomIntegerGenerator.java
@@ -1,13 +1,0 @@
-package dto.definition.property.value.generator.impl.random.impl.numeric;
-
-public class RandomIntegerGenerator extends AbstractNumericRandomGenerator<Integer> {
-
-    public RandomIntegerGenerator(Integer from, Integer to) {
-        super(from, to);
-    }
-
-    @Override
-    public Integer generateValue() {
-        return from + random.nextInt(to - from);
-    }
-}
