+++ DTO/src/dto/impl/DTOMenu3ForUiContextImpl.java
@@ -1,0 +1,5 @@
+package dto.impl;
+
+public class DTOMenu3ForUiContextImpl {
+    private Context context;
+}
