+++ UI/src/ui/Menu3.java
@@ -1,10 +1,15 @@
 package ui;
 
-import dto.api.DTOMenu3ForUi;
-import dto.definition.property.api.PropertyDefinitionDTO;
+import dto.api.DTOMenu3ForUiEVD;
+import dto.definition.property.definition.api.PropertyDefinitionDTO;
+import dto.definition.property.instance.api.PropertyInstanceDTO;
 import dto.definition.rule.enums.Type;
 import system.engine.api.SystemEngineAccess;
-import system.engine.world.execution.instance.property.api.PropertyInstance;
+import system.engine.world.execution.instance.enitty.api.EntityInstance;
+import system.engine.world.rule.action.api.Action;
+import system.engine.world.rule.api.Rule;
+import system.engine.world.rule.context.Context;
+import system.engine.world.rule.context.ContextImpl;
 import ui.dto.creation.CreateDTOMenu3ForSE;
 
 import java.io.BufferedReader;
@@ -12,12 +17,12 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.List;
 import java.util.ArrayList;
+import java.util.stream.Stream;
 
 public class Menu3 {
 
     public void executeSimulation(SystemEngineAccess systemEngineAccess){
-        DTOMenu3ForUi dtoMenu3 = systemEngineAccess.getDataForMenu3FromSE();
-        int countEnvVar = 0;
+        DTOMenu3ForUiEVD dtoMenu3 = systemEngineAccess.getEVDForMenu3FromSE();
         List<Object> initValues = new ArrayList<>();
 
         System.out.println("Here is the list of environment variables.\n" +
@@ -27,16 +32,40 @@ public class Menu3 {
 
         printEnvironmentVarsDataAndCollectValueFromUser(dtoMenu3, initValues);
         systemEngineAccess.updateEnvironmentVarDefinition(new CreateDTOMenu3ForSE().getDataForMenu3(initValues));
-        printEnvironmentVarsDataAfterGeneration(systemEngineAccess.getEnvironmentVarInstances());
+        printEnvironmentVarsDataAfterGeneration(systemEngineAccess.getEVIForMenu3FromSE().getEnvironmentVars());
         systemEngineAccess.addWorldInstance();
         runSimulation(systemEngineAccess);
     }
 
     public void runSimulation(SystemEngineAccess systemEngineAccess){    //on last index at world instances list
-            //iterations by ticks
+            int tick = 0;
+            int seconds = 0;
+            int numOfTicksToRun = systemEngineAccess.getNumOfTicksToRun();
+            int numOfSecondsToRun = systemEngineAccess.getNumOfSecondsToRun();
+
+            List<EntityInstance> entitiesToKill = new ArrayList<>();
+
+            List<Action> actionsList = new ArrayList<>();
+            Stream<Action> actionsStream;
+
+
+            while (tick<= numOfTicksToRun && seconds<=numOfSecondsToRun){
+                for(Rule rule : systemEngineAccess.getActiveRules(tick)){
+                    actionsList.addAll(rule.getActionsToPerform());
+                 }
+                actionsStream = Stream.of((Action) actionsList);
+
+                for(EntityInstance entityInstance : systemEngineAccess.getAllInstancesOfLastWorldInstance()){
+                    for (Action action : actionsList){
+                        Context context = new ContextImpl(entityInstance, systemEngineAccess.getEntityInstanceManagerOfLastWorldInstance(),
+                                systemEngineAccess.getEnvVariablesInstanceManager());
+                        action.executeAction(context);
+                    }
+                }
+            }
     }
 
-    private void printEnvironmentVarsDataAndCollectValueFromUser(DTOMenu3ForUi dtoMenu3, List<Object> initValues){
+    private void printEnvironmentVarsDataAndCollectValueFromUser(DTOMenu3ForUiEVD dtoMenu3, List<Object> initValues){
         int countEnvVar = 0;
 
         for(PropertyDefinitionDTO environmentVar : dtoMenu3.getEnvironmentVars()){
@@ -52,11 +81,11 @@ public class Menu3 {
             }
         }
     }
-    private void printEnvironmentVarsDataAfterGeneration(List<PropertyInstance> envVars) {
+    private void printEnvironmentVarsDataAfterGeneration(List<PropertyInstanceDTO> envVars) {
         int countEnvVar = 0;
 
         System.out.println("Here is the list of all environment variables after generation:");
-        for(PropertyInstance environmentVar : envVars){
+        for(PropertyInstanceDTO environmentVar : envVars){
             countEnvVar++;
             System.out.println("#" + countEnvVar + " name: " + environmentVar.getPropertyDefinition().getUniqueName());
             System.out.println("   " + "value: " + environmentVar.getValue().toString());
