+++ systemEngine/src/dto/creation/CreateDTOMenu2ForUi.java
@@ -3,12 +3,12 @@ package dto.creation;
 import dto.api.DTOMenu2ForUi;
 import dto.definition.entity.api.EntityDefinitionDTO;
 import dto.definition.entity.impl.EntityDefinitionDTOImpl;
-import dto.definition.property.api.PropertyDefinitionDTO;
-import dto.definition.property.impl.BooleanPropertyDefinitionDTO;
-import dto.definition.property.impl.FloatPropertyDefinitionDTO;
-import dto.definition.property.impl.IntegerPropertyDefinitionDTO;
-import dto.definition.property.impl.StringPropertyDefinitionDTO;
-import dto.definition.property.value.generator.api.ValueGenerator;
+import dto.definition.property.definition.api.PropertyDefinitionDTO;
+import dto.definition.property.definition.impl.BooleanPropertyDefinitionDTO;
+import dto.definition.property.definition.impl.FloatPropertyDefinitionDTO;
+import dto.definition.property.definition.impl.IntegerPropertyDefinitionDTO;
+import dto.definition.property.definition.impl.StringPropertyDefinitionDTO;
+import dto.definition.property.definition.value.generator.api.ValueGenerator;
 import dto.definition.rule.activation.impl.ActivationDTOImpl;
 import dto.definition.rule.api.RuleDTO;
 import dto.definition.rule.impl.RuleDTOImpl;
