--- systemEngine/src/dto/creation/CreateDTOMenu3ForUi.java
@@ -1,41 +1,0 @@
-package dto.creation;
-
-import dto.api.DTOMenu3ForUi;
-import dto.definition.property.api.PropertyDefinitionDTO;
-import dto.definition.property.impl.BooleanPropertyDefinitionDTO;
-import dto.definition.property.impl.FloatPropertyDefinitionDTO;
-import dto.definition.property.impl.IntegerPropertyDefinitionDTO;
-import dto.definition.property.impl.StringPropertyDefinitionDTO;
-import dto.definition.property.value.generator.api.ValueGenerator;
-import dto.impl.DTOMenu3ForUiImpl;
-import system.engine.world.api.WorldDefinition;
-import system.engine.world.definition.property.api.PropertyDefinition;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class CreateDTOMenu3ForUi  {
-
-    public DTOMenu3ForUi getDataForMenu3(WorldDefinition worldDefinition) {
-        List<PropertyDefinitionDTO> environmentVars= new ArrayList<>();
-
-        for(PropertyDefinition environmentVar : worldDefinition.getEnvVariablesDefinitionManager().getEnvVariables()){
-            environmentVars.add(createEnvironmentVarDTO(environmentVar));
-        }
-        return new DTOMenu3ForUiImpl(environmentVars);
-    }
-
-    private PropertyDefinitionDTO createEnvironmentVarDTO(PropertyDefinition environmentVar){
-        switch (environmentVar.getType()){
-            case DECIMAL:
-                return new IntegerPropertyDefinitionDTO(environmentVar.getUniqueName(), (ValueGenerator<Integer>) environmentVar.getValueGenerator());
-            case FLOAT:
-                return new FloatPropertyDefinitionDTO(environmentVar.getUniqueName(), (ValueGenerator<Float>) environmentVar.getValueGenerator());
-            case STRING:
-                return new StringPropertyDefinitionDTO(environmentVar.getUniqueName(), (ValueGenerator<String>) environmentVar.getValueGenerator());
-            case BOOLEAN:
-                return new BooleanPropertyDefinitionDTO(environmentVar.getUniqueName(), (ValueGenerator<Boolean>) environmentVar.getValueGenerator());
-        }
-        return null;
-    }
-}
